refactor(frontend): some improvements to semester parsing and normalization
